Load .env variables in config/config.exs so it works for all environments
