+++ .env.sample
@@ -1,4 +1,5 @@
 PGUSER=username
 PGPASSWORD=password
-PGDATABASE=jeopardixir_dev
+PGDATABASE_DEV=jeopardixir_dev
+PGDATABASE_TEST=jeopardixir_test
 PGHOST=localhost
